# Use Expression to create instance for better performance